requirements.txt
python-telegram-bot~=20.7
python-dotenv~=1.1.0
pytest~=8.4.1